feat: 'useCurrentLocation 커스텀 훅 구현'
weekly bar menu에서 필요한 current location의 코드 재사용을 위한 useCurrentLocation 훅 생성.
연관된 weather data 관련 코드 api 폴더 내 파일 생성하여 이동.
weather data 관련 interface를 2개 이상 컴포넌트에서 재사용함에 따라 type.d.ts 파일 생성 및 코드 이동해 전역 사용.
'sky is clear': '맑음' -> 'clear sky': '맑음'으로 기존 날씨 번역 객체 수정함.

Related to: #42
