+++ src/components/template/MainTemplate.tsx
@@ -3,43 +3,30 @@ import SliderMolecule from '../molecule/SliderMolecule';
 import WeatherOrganism from '../organism/WeatherOrganism';
 import PopularLocationOrganism from '../organism/PopularLocationOrganism';
 import CommunityOrganism from '../organism/CommunityOrganism';
-import { useState, useEffect } from 'react';
-
-interface WeatherData {
-  name: string;
-  main: {
-    temp: number;
-  };
-  weather: {
-    description: string;
-  }[];
-}
+import { useEffect, useState } from 'react';
+import { useCurrentLocation } from '../../hooks/useCurrentLocation';
+import { getWeatherData } from '../../api/weather';
 
 const MainTemplate = () => {
-  const [weather, setWeather] = useState<WeatherData | null>(null);
+  const [weather, setWeather] = useState<WeatherData | null>(null); // 날씨 데이터
+  const { lat, lon } = useCurrentLocation(); // 현재 위치 받아오는 custom hook
 
-  // 현재 위치 가져오기
-  const getCurrentLocation = () => {
-    navigator.geolocation.getCurrentPosition((position) => {
-      let lat = position.coords.latitude;
-      let lon = position.coords.longitude;
-      // console.log(lat, lon);
-      getWeatherByCurrentLocation(lat, lon);
-    });
+  // 현재 위치 기반 날씨 가져오는 함수
+  const getWeatherDataByCurrentLocation = async (lat: number, lon: number) => {
+    try {
+      const data = await getWeatherData(lat, lon);
+      setWeather(data);
+    } catch (error) {
+      console.error(`데이터를 불러오지 못했습니다. : ${error}`);
+    }
   };
 
-  // 현재 위치의 날씨 가져오기
-  const getWeatherByCurrentLocation = async (lat: number, lon: number) => {
-    let url = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=576217254852d28ff29e275ca843c5fe&units=metric`;
-    let response = await fetch(url);
-    let data = await response.json();
-    // console.log(data);
-    setWeather(data);
-  };
-
+  // 현재 위치 기반 날씨 가져오는 함수에 arguments로 현재 위치 삽입해 현재 위치 날씨 가져오기
   useEffect(() => {
-    getCurrentLocation();
-  }, []);
+    if (lat !== null && lon !== null) {
+      getWeatherDataByCurrentLocation(lat, lon);
+    }
+  }, [lat, lon]);
 
   return (
     <>
